+++ nf_core/modules/update.py
@@ -1,9 +1,13 @@
-import os
-import shutil
-import questionary
-import logging
-import tempfile
+import copy
 import difflib
+import enum
+import json
+import logging
+import os
+import questionary
+import shutil
+import tempfile
+from questionary import question
 from rich.console import Console
 from rich.syntax import Syntax
 
@@ -18,13 +22,14 @@ log = logging.getLogger(__name__)
 
 
 class ModuleUpdate(ModuleCommand):
-    def __init__(self, pipeline_dir, force=False, prompt=False, sha=None, update_all=False, diff=False):
+    def __init__(self, pipeline_dir, force=False, prompt=False, sha=None, update_all=False, save_diff_fn=None):
         super().__init__(pipeline_dir)
         self.force = force
         self.prompt = prompt
         self.sha = sha
         self.update_all = update_all
-        self.diff = diff
+        self.show_diff = False
+        self.save_diff_fn = save_diff_fn
 
     def update(self, module):
         if self.repo_type == "modules":
@@ -157,6 +162,7 @@ class ModuleUpdate(ModuleCommand):
                         repos_mods_shas[repo_name].append((module, custom_sha))
                 else:
                     skipped_repos.append(repo_name)
+
             if skipped_repos:
                 skipped_str = "', '".join(skipped_repos)
                 log.info(f"Skipping modules in repositor{'y' if len(skipped_repos) == 1 else 'ies'}: '{skipped_str}'")
@@ -177,12 +183,49 @@ class ModuleUpdate(ModuleCommand):
 
         # Load 'modules.json'
         modules_json = self.load_modules_json()
+        old_modules_json = copy.deepcopy(modules_json)  # Deep copy to avoid mutability
         if not modules_json:
             return False
 
+        # Ask if we should show the diffs (unless a filename was already given on the command line)
+        if not self.save_diff_fn:
+            diff_type = questionary.select(
+                "Do you want to view diffs of the proposed changes?",
+                choices=[
+                    {"name": "No previews, just update everything", "value": 0},
+                    {"name": "Preview diff in terminal, choose whether to update files", "value": 1},
+                    {"name": "Just write diffs to a patch file", "value": 2},
+                ],
+                style=nf_core.utils.nfcore_question_style,
+            ).unsafe_ask()
+
+            self.show_diff = diff_type == 1
+            self.save_diff_fn = diff_type == 2
+
+        # Set up file to save diff
+        if self.save_diff_fn:  # True or a string
+            # From questionary - no filename yet
+            if self.save_diff_fn is True:
+                self.save_diff_fn = questionary.text(
+                    "Enter the filename: ", style=nf_core.utils.nfcore_question_style
+                ).unsafe_ask()
+            # Check if filename already exists (questionary or cli)
+            while os.path.exists(self.save_diff_fn):
+                if questionary.confirm(f"'{self.save_diff_fn}' exists. Remove file?").unsafe_ask():
+                    os.remove(self.save_diff_fn)
+                    break
+                self.save_diff_fn = questionary.text(
+                    f"Enter a new filename: ",
+                    style=nf_core.utils.nfcore_question_style,
+                ).unsafe_ask()
+
         exit_value = True
         for modules_repo, module, sha in repos_mods_shas:
-            dry_run = self.diff
+
+            # Are we updating the files in place or not?
+            dry_run = self.show_diff or self.save_diff_fn
+
+            # Check if the module we've been asked to update actually exists
             if not module_exist_in_repo(module, modules_repo):
                 warn_msg = f"Module '{module}' not found in remote '{modules_repo.name}' ({modules_repo.branch})"
                 if self.update_all:
@@ -252,61 +295,160 @@ class ModuleUpdate(ModuleCommand):
                 continue
 
             if dry_run:
-                console = Console(force_terminal=nf_core.utils.rich_force_colors())
-                files = os.listdir(os.path.join(*install_folder, module))
-                temp_folder = os.path.join(*install_folder, module)
-                log.info(
-                    f"Changes in module '{module}' between ({current_entry['git_sha'] if current_entry is not None else '?'}) and ({version if version is not None else 'latest'})"
-                )
 
+                class DiffEnum(enum.Enum):
+                    """
+                    Enumeration for keeping track of
+                    the diff status of a pair of files
+                    """
+
+                    UNCHANGED = enum.auto()
+                    CHANGED = enum.auto()
+                    CREATED = enum.auto()
+                    REMOVED = enum.auto()
+
+                diffs = {}
+
+                # Get all unique filenames in the two folders.
+                # `dict.fromkeys()` is used instead of `set()` to preserve order
+                files = dict.fromkeys(os.listdir(os.path.join(*install_folder, module)))
+                files.update(dict.fromkeys(os.listdir(module_dir)))
+                files = list(files)
+
+                temp_folder = os.path.join(*install_folder, module)
+
+                # Loop through all the module files and compute their diffs if needed
                 for file in files:
                     temp_path = os.path.join(temp_folder, file)
                     curr_path = os.path.join(module_dir, file)
-                    if os.path.exists(temp_path) and os.path.exists(curr_path):
+                    if os.path.exists(temp_path) and os.path.exists(curr_path) and os.path.isfile(temp_path):
                         with open(temp_path, "r") as fh:
                             new_lines = fh.readlines()
                         with open(curr_path, "r") as fh:
                             old_lines = fh.readlines()
+
                         if new_lines == old_lines:
                             # The files are identical
-                            log.info(f"'{os.path.join(module, file)}' is unchanged")
+                            diffs[file] = (DiffEnum.UNCHANGED, ())
                         else:
-                            log.info(f"Changes in '{os.path.join(module, file)}':")
                             # Compute the diff
                             diff = difflib.unified_diff(
                                 old_lines,
                                 new_lines,
-                                fromfile=f"{os.path.join(module, file)} (installed)",
-                                tofile=f"{os.path.join(module, file)} (new)",
+                                fromfile=os.path.join(module_dir, file),
+                                tofile=os.path.join(module_dir, file),
                             )
+                            diffs[file] = (DiffEnum.CHANGED, diff)
 
+                    elif os.path.exists(temp_path):
+                        # The file was created
+                        diffs[file] = (DiffEnum.CREATED, ())
+
+                    elif os.path.exists(curr_path):
+                        # The file was removed
+                        diffs[file] = (DiffEnum.REMOVED, ())
+
+                if self.save_diff_fn:
+                    log.info(f"Writing diff of '{module}' to '{self.save_diff_fn}'")
+                    with open(self.save_diff_fn, "a") as fh:
+                        fh.write(
+                            f"Changes in module '{module}' between ({current_entry['git_sha'] if current_entry is not None else '?'}) and ({version if version is not None else 'latest'})\n"
+                        )
+
+                        for file, d in diffs.items():
+                            diff_status, diff = d
+                            if diff_status == DiffEnum.UNCHANGED:
+                                # The files are identical
+                                fh.write(f"'{os.path.join(module_dir, file)}' is unchanged\n")
+
+                            elif diff_status == DiffEnum.CREATED:
+                                # The file was created between the commits
+                                fh.write(f"'{os.path.join(module_dir, file)}' was created\n")
+
+                            elif diff_status == DiffEnum.REMOVED:
+                                # The file was removed between the commits
+                                fh.write(f"'{os.path.join(module_dir, file)}' was removed\n")
+
+                            else:
+                                # The file has changed
+                                fh.write(f"Changes in '{os.path.join(module_dir, file)}':\n")
+                                # Write the diff lines to the file
+                                for line in diff:
+                                    fh.write(line)
+                                fh.write("\n")
+
+                        fh.write("*" * 60 + "\n")
+                elif self.show_diff:
+                    console = Console(force_terminal=nf_core.utils.rich_force_colors())
+                    log.info(
+                        f"Changes in module '{module}' between ({current_entry['git_sha'] if current_entry is not None else '?'}) and ({version if version is not None else 'latest'})"
+                    )
+
+                    for file, d in diffs.items():
+                        diff_status, diff = d
+                        if diff_status == DiffEnum.UNCHANGED:
+                            # The files are identical
+                            log.info(f"'{os.path.join(module, file)}' is unchanged")
+                        elif diff_status == DiffEnum.CREATED:
+                            # The file was created between the commits
+                            log.info(f"'{os.path.join(module, file)}' was created")
+                        elif diff_status == DiffEnum.REMOVED:
+                            # The file was removed between the commits
+                            log.info(f"'{os.path.join(module, file)}' was removed")
+                        else:
+                            # The file has changed
+                            log.info(f"Changes in '{os.path.join(module, file)}':")
                             # Pretty print the diff using the pygments diff lexer
                             console.print(Syntax("".join(diff), "diff", theme="ansi_light"))
 
-                    elif os.path.exists(temp_path):
-                        # The file was created between the commits
-                        log.info(f"Created file '{file}'")
+                    # Ask the user if they want to install the module
+                    dry_run = not questionary.confirm(
+                        f"Update module '{module}'?", default=False, style=nf_core.utils.nfcore_question_style
+                    ).unsafe_ask()
+                    if not dry_run:
+                        # The new module files are already installed.
+                        # We just need to clear the directory and move the
+                        # new files from the temporary directory
+                        self.clear_module_dir(module, module_dir)
+                        os.makedirs(module_dir)
+                        for file in files:
+                            path = os.path.join(temp_folder, file)
+                            if os.path.exists(path):
+                                shutil.move(path, os.path.join(module_dir, file))
+                        log.info(f"Updating '{modules_repo.name}/{module}'")
+                        log.debug(f"Updating module '{module}' to {version} from {modules_repo.name}")
 
-                    elif os.path.exists(curr_path):
-                        # The file was removed between the commits
-                        log.info(f"Removed file '{file}'")
-
-                # Ask the user if they want to install the module
-                dry_run = not questionary.confirm("Update module?", default=False).unsafe_ask()
-                if not dry_run:
-                    # The new module files are already installed
-                    # we just need to clear the directory and move the
-                    # new files from the temporary directory
-                    self.clear_module_dir(module, module_dir)
-                    os.mkdir(module_dir)
-                    for file in files:
-                        path = os.path.join(temp_folder, file)
-                        if os.path.exists(path):
-                            shutil.move(path, os.path.join(module_dir, file))
-                    log.info(f"Updating '{modules_repo.name}/{module}'")
-                    log.debug(f"Updating module '{module}' to {version} from {modules_repo.name}")
-
+            # Update modules.json with newly installed module
             if not dry_run:
-                # Update module.json with newly installed module
                 self.update_modules_json(modules_json, modules_repo.name, module, version)
+
+            # Don't save to a file, just iteratively update the variable
+            else:
+                modules_json = self.update_modules_json(
+                    modules_json, modules_repo.name, module, version, write_file=False
+                )
+
+        if self.save_diff_fn:
+            # Compare the new modules.json and build a diff
+            modules_json_diff = difflib.unified_diff(
+                json.dumps(old_modules_json, indent=4).splitlines(keepends=True),
+                json.dumps(modules_json, indent=4).splitlines(keepends=True),
+                fromfile=os.path.join(self.dir, "modules.json"),
+                tofile=os.path.join(self.dir, "modules.json"),
+            )
+
+            # Save diff for modules.json to file
+            with open(self.save_diff_fn, "a") as fh:
+                fh.write(f"Changes in './modules.json'\n")
+                for line in modules_json_diff:
+                    fh.write(line)
+                fh.write("*" * 60 + "\n")
+
+        log.info("Updates complete :sparkles:")
+
+        if self.save_diff_fn:
+            log.info(
+                f"[bold magenta italic] TIP! [/] If you are happy with the changes in '{self.save_diff_fn}', you can apply them by running the command :point_right:  [bold magenta italic]git apply {self.save_diff_fn}"
+            )
+
         return exit_value